fix(post): move global loading effect to a component to resolve issue with /post/[slug] URLs (#653)
